Ensure that all native callbacks are setup always
