+++ Realm/Realm/Handles/AppHandle.cs
@@ -47,12 +47,15 @@ namespace Realms.Sync
             [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
             public unsafe delegate void BsonCallback(IntPtr tcs_ptr, BsonPayload response, AppError error);
 
+            [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+            public delegate void ApiKeysCallback(IntPtr tcs_ptr, /* UserApiKey[] */ IntPtr api_keys, int api_keys_len, AppError error);
+
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_app_initialize", CallingConvention = CallingConvention.Cdecl)]
             public static extern IntPtr initialize(
                 [MarshalAs(UnmanagedType.LPWStr)] string platform, IntPtr platform_len,
                 [MarshalAs(UnmanagedType.LPWStr)] string platform_version, IntPtr platform_version_len,
                 [MarshalAs(UnmanagedType.LPWStr)] string sdk_version, IntPtr sdk_version_len,
-                UserCallback user_callback, VoidTaskCallback void_callback, BsonCallback bson_callback, LogMessageCallback log_message_callback);
+                UserCallback user_callback, VoidTaskCallback void_callback, BsonCallback bson_callback, LogMessageCallback log_message_callback, ApiKeysCallback api_keys_callback);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_app_create", CallingConvention = CallingConvention.Cdecl)]
             public static extern IntPtr create_app(Native.AppConfiguration app_config, byte[] encryptionKey, out NativeException ex);
@@ -143,17 +146,21 @@ namespace Realms.Sync
         static unsafe AppHandle()
         {
             NativeCommon.Initialize();
-            SessionHandle.InstallCallbacks();
+        }
 
+        public static void Initialize()
+        {
             NativeMethods.LogMessageCallback logMessage = HandleLogMessage;
             NativeMethods.UserCallback userLogin = HandleUserCallback;
             NativeMethods.VoidTaskCallback taskCallback = HandleTaskCompletion;
             NativeMethods.BsonCallback bsonCallback = HandleBsonCallback;
+            NativeMethods.ApiKeysCallback apiKeysCallback = HandleApiKeysCallback;
 
             GCHandle.Alloc(logMessage);
             GCHandle.Alloc(userLogin);
             GCHandle.Alloc(taskCallback);
             GCHandle.Alloc(bsonCallback);
+            GCHandle.Alloc(apiKeysCallback);
 
             //// This is a hack due to a mixup of what OS uses as platform/SDK and what is displayed in the UI.
             //// The original code is below:
@@ -183,9 +190,7 @@ namespace Realms.Sync
                 platform, platform.IntPtrLength(),
                 platformVersion, platformVersion.IntPtrLength(),
                 sdkVersion, sdkVersion.IntPtrLength(),
-                userLogin, taskCallback, bsonCallback, logMessage);
-
-            HttpClientTransport.Install();
+                userLogin, taskCallback, bsonCallback, logMessage, apiKeysCallback);
         }
 
         internal AppHandle(IntPtr handle) : base(null, handle)
@@ -364,6 +369,34 @@ namespace Realms.Sync
                 if (error.is_null)
                 {
                     tcs.TrySetResult(response);
+                }
+                else
+                {
+                    tcs.TrySetException(new AppException(error));
+                }
+            }
+            finally
+            {
+                tcsHandle.Free();
+            }
+        }
+
+        [MonoPInvokeCallback(typeof(NativeMethods.ApiKeysCallback))]
+        private static unsafe void HandleApiKeysCallback(IntPtr tcs_ptr, IntPtr api_keys, int api_keys_len, AppError error)
+        {
+            var tcsHandle = GCHandle.FromIntPtr(tcs_ptr);
+            try
+            {
+                var tcs = (TaskCompletionSource<UserApiKey[]>)tcsHandle.Target;
+                if (error.is_null)
+                {
+                    var result = new UserApiKey[api_keys_len];
+                    for (var i = 0; i < api_keys_len; i++)
+                    {
+                        result[i] = Marshal.PtrToStructure<UserApiKey>(IntPtr.Add(api_keys, i * UserApiKey.Size));
+                    }
+
+                    tcs.TrySetResult(result);
                 }
                 else
                 {
