+++ Realm/Realm/Handles/SessionHandle.cs
@@ -32,7 +32,7 @@ namespace Realms.Sync
         private static class NativeMethods
         {
 #pragma warning disable IDE1006 // Naming Styles
-#pragma warning disable SA1121 // Use built-in type alias
+
             [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
             public unsafe delegate void SessionErrorCallback(IntPtr session_handle_ptr, ErrorCode error_code, byte* message_buf, IntPtr message_len, IntPtr user_info_pairs, int user_info_pairs_len, [MarshalAs(UnmanagedType.U1)] bool is_client_reset);
 
@@ -87,12 +87,6 @@ namespace Realms.Sync
             public static extern void start(SessionHandle session, out NativeException ex);
 
 #pragma warning restore IDE1006 // Naming Styles
-#pragma warning restore SA1121 // Use built-in type alias
-        }
-
-        static SessionHandle()
-        {
-            NativeCommon.Initialize();
         }
 
         [Preserve]
