+++ Realm/Realm/Handles/SyncUserHandle.cs
@@ -21,8 +21,6 @@ using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using MongoDB.Bson;
 using Realms.Native;
-using Realms.Sync.Exceptions;
-using Realms.Sync.Native;
 
 namespace Realms.Sync
 {
@@ -31,10 +29,6 @@ namespace Realms.Sync
         private static class NativeMethods
         {
 #pragma warning disable IDE1006 // Naming Styles
-#pragma warning disable SA1121 // Use built-in type alias
-
-            [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-            public delegate void ApiKeysCallback(IntPtr tcs_ptr, /* UserApiKey[] */ IntPtr api_keys, int api_keys_len, AppError error);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "realm_syncuser_get_id", CallingConvention = CallingConvention.Cdecl)]
             public static extern IntPtr get_user_id(SyncUserHandle user, IntPtr buffer, IntPtr buffer_length, out NativeException ex);
@@ -74,9 +68,6 @@ namespace Realms.Sync
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "realm_syncuser_destroy", CallingConvention = CallingConvention.Cdecl)]
             public static extern void destroy(IntPtr syncuserHandle);
-
-            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "realm_sync_user_initialize", CallingConvention = CallingConvention.Cdecl)]
-            public static extern IntPtr initialize(ApiKeysCallback api_keys_callback);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "realm_syncuser_call_function", CallingConvention = CallingConvention.Cdecl)]
             public static extern void call_function(SyncUserHandle handle, AppHandle app,
@@ -130,18 +121,6 @@ namespace Realms.Sync
             #endregion
 
 #pragma warning restore IDE1006 // Naming Styles
-#pragma warning restore SA1121 // Use built-in type alias
-        }
-
-        static unsafe SyncUserHandle()
-        {
-            NativeCommon.Initialize();
-
-            NativeMethods.ApiKeysCallback apiKeysCallback = HandleApiKeysCallback;
-
-            GCHandle.Alloc(apiKeysCallback);
-
-            NativeMethods.initialize(apiKeysCallback);
         }
 
         [Preserve]
@@ -340,34 +319,6 @@ namespace Realms.Sync
         protected override void Unbind()
         {
             NativeMethods.destroy(handle);
-        }
-
-        [MonoPInvokeCallback(typeof(NativeMethods.ApiKeysCallback))]
-        private static unsafe void HandleApiKeysCallback(IntPtr tcs_ptr, IntPtr api_keys, int api_keys_len, AppError error)
-        {
-            var tcsHandle = GCHandle.FromIntPtr(tcs_ptr);
-            try
-            {
-                var tcs = (TaskCompletionSource<UserApiKey[]>)tcsHandle.Target;
-                if (error.is_null)
-                {
-                    var result = new UserApiKey[api_keys_len];
-                    for (var i = 0; i < api_keys_len; i++)
-                    {
-                        result[i] = Marshal.PtrToStructure<UserApiKey>(IntPtr.Add(api_keys, i * UserApiKey.Size));
-                    }
-
-                    tcs.TrySetResult(result);
-                }
-                else
-                {
-                    tcs.TrySetException(new AppException(error));
-                }
-            }
-            finally
-            {
-                tcsHandle.Free();
-            }
         }
     }
 }
