+++ Realm/Realm/Native/NativeCommon.cs
@@ -22,6 +22,8 @@ using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Threading;
+using Realms.Native;
+using Realms.Sync;
 
 namespace Realms
 {
@@ -49,6 +51,13 @@ namespace Realms
                     // This is the path in the Unity package - it is what the Editor uses.
                     AddWindowsWrappersToPath("Windows", isUnityTarget: true);
                 }
+
+                SynchronizationContextScheduler.Install();
+                SharedRealmHandle.Initialize();
+                SessionHandle.InstallCallbacks();
+                HttpClientTransport.Install();
+                AppHandle.Initialize();
+
             }
         }
